+++ common/src/main/java/immersive_armors/client/render/entity/model/VerticalHeadModel.java
@@ -6,11 +6,15 @@ public class VerticalHeadModel extends DecoHeadModel {
     private final ModelPart part;
 
     public VerticalHeadModel() {
+        this(0.0f, 0.0f, 0.0f);
+    }
+
+    public VerticalHeadModel(float x, float y, float z) {
         super();
 
         part = new ModelPart(64, 16, 0, 0);
         part.addCuboid(0.0F, -17.0F, -10.0F, 0.0F, 12.0F, 20.0F, 0.0f);
-        part.setPivot(0.0F, 0.0F, 0.0F);
+        part.setPivot(x, y, z);
     }
 
     @Override