+++ common/src/main/java/immersive_armors/client/render/entity/piece/Piece.java
@@ -2,6 +2,7 @@ package immersive_armors.client.render.entity.piece;
 
 import immersive_armors.item.ArmorPiece;
 import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.render.entity.model.BipedEntityModel;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ItemStack;
@@ -11,5 +12,5 @@ public abstract class Piece {
 
     }
 
-    public abstract void render(MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, LivingEntity entity, ItemStack itemStack, float tickDelta, ArmorPiece piece);
+    public abstract void render(MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, LivingEntity entity, ItemStack itemStack, float tickDelta, ArmorPiece piece, BipedEntityModel<LivingEntity> bipedEntityModel);
 }